MORE!

+++ typings/geopy/distance.pyi
@@ -1,4 +1,4 @@
 class distance:  # noqa: N801
-    km: int
+    km: float
 
     def __init__(self, *args: tuple[float, float]) -> None: ...
